rebar3_erlydtl_plugin (rebar3_erlydtl_plugin v0.1.1)

The rebar_erlydtl_compiler module is a plugin for rebar that compiles ErlyDTL templates. By default, it compiles all priv/templates/*.dtl to ebin/*_dtl.beam.

Configuration options should be placed in rebar.config under erlydtl_opts. It can be a list of name-value tuples or a list of lists of name-value tuples if you have multiple template directories that need to have different settings (see example below).
Available options include:

doc_root: where to find templates to compile priv/templates by default

out_dir: where to put compiled template beam files ebin by default

source_ext: the file extension the template sources have .dtl by default

module_ext: characters to append to the template's module name _dtl by default

recursive: boolean that determines if doc_root(s) need to be scanned recursively for matching template file names (default: true).

For example, if you had:
    /t_src/
           base.html
           foo.html
And you wanted them compiled to:
    /priv/
          base.beam
          foo.beam
You would add to your rebar.config:
    {erlydtl_opts, [
                {doc_root,   "t_src"},
                {out_dir,    "priv"},
                {source_ext, ".html"},
                {module_ext, ""}
               ]}.
The default settings are the equivalent of:
    {erlydtl_opts, [
                {doc_root,   "priv/templates"},
                {out_dir,    "ebin"},
                {source_ext, ".dtl"},
                {module_ext, "_dtl"}
               ]}.
The following example will compile the following templates: src/*.dtl files into ebin/*_dtl.beam and priv/templates/*.html into ebin/*.beam. Note that any tuple option (such as out_dir) in the outer list is added to each inner list:
    {erlydtl_opts, [
       {out_dir, "ebin"},
       {recursive, false},
       [
           {doc_root, "src"}, {module_ext, "_dtl"}
       ],
       [
           {doc_root, "priv/templates"}, {module_ext, ""}, {source_ext, ".html"}
       ]
    ]}.

Functions

do(State)

format_error(Reason)

-spec format_error(any()) -> iolist().

init(State)

-spec init(rebar_state:t()) -> {ok, rebar_state:t()}.